Initial work on integrating management service
This is the first commit of the management service integration. It is not complete yet, but it is touching on most of the files that will be affected by the integration.

We may want to organize the code a bit better, but I wanted to get this out there so we can start discussing it.